Implement buying feature

+++ src/com/gaji/mini/member/model/vo/Member.java
@@ -38,7 +38,7 @@ public abstract class Member implements Serializable {
 		this.pw = pw;
 	}
 
-	public long getMoney() {
+	public int getMoney() {
 		return money;
 	}
